+++ TombEngine/Scripting/Internal/TEN/Flow/FlowHandler.h
@@ -64,6 +64,8 @@ public:
 	int			GetNumLevels() const;
 	void		EndLevel(std::optional<int> nextLevel, std::optional<int> startPosIndex);
 	GameStatus	GetGameStatus();
+	FreezeMode	GetFreezeMode();
+	void		SetFreezeMode(FreezeMode mode);
 	void		FlipMap(int group);
 	bool		GetFlipMapStatus(std::optional<int> group);
 	void		SaveGame(int slot);